requirements.txt
pyjwt~=2.8.0
pydantic~=1.10.12
fastapi~=0.103.0
uvicorn~=0.20.0